--- src/PAGE-2020.jl
@@ -1,5 +1,0 @@
-module PAGE2020
-
-include("main_model.jl")
-
-end
