+++ src/components/CO2cycle.jl
@@ -1,5 +1,3 @@
-using Mimi
-
 @defcomp CO2Cycle begin
     # Permafrost outputs
     permte0_permafrostemissions0 = Parameter(unit="Mtonne")
