+++ src/components/PAGE09SeaLevelRise.jl
@@ -1,5 +1,3 @@
-using Mimi
-
 @defcomp PAGE09SeaLevelRise begin
 
   # Parameters
