+++ src/compute_scc.jl
@@ -1,5 +1,3 @@
-using Mimi
-
 page_years = [2020, 2030, 2040, 2050, 2075, 2100, 2150, 2200, 2250, 2300]
 page_year_0 = 2015
