+++ src/utils/mctools.jl
@@ -1,5 +1,4 @@
 using Query
-using FileIO
 
 """
 Create a parameter `component`_`name` with the given value,
