+++ test/test_MarketDamages.jl
@@ -1,6 +1,6 @@
 using Test
 
-=for testscen in 1:2
+for testscen in 1:2
     valdir, scenario, use_permafrost, use_seaice = get_scenario(testscen)
     println(scenario)
