+++ test/test_scc.jl
@@ -30,5 +30,3 @@ checksccs(sccs, [53.903, 129.330, 217.337, 361.739, 886.998, 306.914])
 
 # m = getpage("RCP8.5 & SSP5", true, true, false)
 # sccs = compute_scc_mcs(m, 1000, year=2020)
-
-end
